compile
svn path=/trunk/kdeextragear-1/amarok/; revision=343230

+++ src/contextbrowser.cpp
@@ -392,7 +392,7 @@ void ContextBrowser::showCurrentTrack() //SLOT
                                     "<table width='100%'><tr><td width='20%'><a class='menu' href='fetchcover:%7 @@@ %8'>"
                                     "<img hspace='2' src='%9'></a></td>"
                                     "<td valign='bottom' align='right' width='80%'>" +
-                                    i18n( "Track played 1 time", "Track played %n times", values[4]) + "<br>" +
+                                    i18n( "Track played 1 time", "Track played %n times", values[4].toInt()) + "<br>" +
                                     "amaroKiness: %11" + "<br>" +
                                     i18n( "Last play: %12" ) + "<br>" +
                                     i18n( "First play: %13") + "</i></td></tr>" )
